[Iteration 12] - [Amit] Create Order Screen
This check in has the following changes:
1. The bean transformation to save the order details.
2. The order transformer changes.
3. The location list on the order screen with error handling.
4. the item lookup based on order cost details along with supplier  cost.

+++ src/main/java/com/punj/app/ecommerce/services/impl/OrderServiceImpl.java
@@ -379,7 +379,7 @@ public class OrderServiceImpl implements OrderService {
 			itemStockJournal.setReasonCode(stockReason);
 
 			itemStockJournal.setFunctionality(ServiceConstants.RECEIVE_ORDER_FUNCTIONALITY);
-			itemStockJournal.setQty(orderItem.getDelieveredQty());
+			itemStockJournal.setQty(orderItem.getDelieveredQty().intValue());
 
 			itemStockDetails.add(itemStockJournal);
 		}
